pyproject.toml
[project]
name = "freshmcp"
version = "0.1.0"
description = "A Python-based service that provides a Message Control Protocol (MCP) interface for FreshMCP operations using Azure Cosmos DB and AI Search"
authors = [
    {name = "Fresh AI Team"}
]
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "python-dotenv==1.1.0",
    "mcp==1.9.0",
    "azure.identity==1.21.0",
    "sse-starlette==2.2.1",
    "starlette==0.46.0",
    "azure-cosmos==4.9.0",
    "azure-search-documents>=11.5.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["freshmcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"